Added new column (container) in item-page.php to mimic the space used by other sites for ads.
Implemented a Quality of Life change for admin-order-search.php.
The effect is that if a bad search is made, the text still remains in the
search box so the user can see if he entered the wrong info. Also just
looks more professional.

Added item-pages.css to remove formatting code from item-pages.php.